refactor: split WindowManager and DragSnapManager along their seams
WindowManager.swift held three types, which the "one type per file" rule
in AGENTS.md rules out and which is most of why it was the longest file
in the app. FracRect and Preset are now files of their own, and the
Accessibility layer -- the calls that read and write another app's
windows -- is WindowManager+AX beside them. What is left decides where a
window goes, which is the one job the type is named for.

DragSnapManager keeps the drag state machine and the overlays. Working
out which zone the cursor is over moves to DragSnapManager+Zones: the
hovered zone, the span back to where the drag started, the straddle
across a seam, and the edge fallback for a screen with no zone set.

  WindowManager     428 -> 260   (+AX 114, FracRect 26, Preset 47)
  DragSnapManager   351 -> 208   (+Zones 156)

None of this is reachable from the unit suite -- AGENTS.md says so, and
it is why the change is a pure move with no logic touched. Verified by
comparing every non-comment line before and after: nothing lost, and the
only additions are the new files' headers. Some members lost `private`
because Swift lets an extension see one only in its declaring file; each
new file says so at the top.

Both come out of scripts/line-limit-baseline, which now holds no
App/Sources file over the limit except GrabMove, UpdateManager and
WorkspaceLauncher.
