+++ App/Sources/plonk/WindowManager.swift
@@ -5,75 +5,11 @@ import ApplicationServices
 // of the primary screen, y grows downward. Fractional frames are relative to
 // a screen's visible area (excludes menu bar and Dock), origin top-left.
 
-struct FracRect {
-    let x, y, w, h: Double
-    init(_ x: Double, _ y: Double, _ w: Double, _ h: Double) {
-        self.x = x; self.y = y; self.w = w; self.h = h
-    }
-
-    /// Parses a `{x,y,w,h}` frame from an API body. Must stay inside the screen
-    /// it is a fraction of; the epsilon absorbs rounding from `get_state`.
-    static func parse(_ value: Any?) -> FracRect? {
-        guard let frame = value as? [String: Any],
-              let x = (frame["x"] as? NSNumber)?.doubleValue,
-              let y = (frame["y"] as? NSNumber)?.doubleValue,
-              let w = (frame["w"] as? NSNumber)?.doubleValue,
-              let h = (frame["h"] as? NSNumber)?.doubleValue,
-              x.isFinite, y.isFinite, w.isFinite, h.isFinite,
-              w > 0, h > 0, x >= 0, y >= 0,
-              x + w <= 1.0001, y + h <= 1.0001 else { return nil }
-        return FracRect(x, y, w, h)
-    }
-}
-
-enum Preset: String, CaseIterable {
-    case leftHalf = "left-half"
-    case rightHalf = "right-half"
-    case topHalf = "top-half"
-    case bottomHalf = "bottom-half"
-    case topLeft = "top-left"
-    case topRight = "top-right"
-    case bottomLeft = "bottom-left"
-    case bottomRight = "bottom-right"
-    case maximize = "maximize"
-    case center = "center"
-
-    var frac: FracRect {
-        switch self {
-        case .leftHalf: return FracRect(0, 0, 0.5, 1)
-        case .rightHalf: return FracRect(0.5, 0, 0.5, 1)
-        case .topHalf: return FracRect(0, 0, 1, 0.5)
-        case .bottomHalf: return FracRect(0, 0.5, 1, 0.5)
-        case .topLeft: return FracRect(0, 0, 0.5, 0.5)
-        case .topRight: return FracRect(0.5, 0, 0.5, 0.5)
-        case .bottomLeft: return FracRect(0, 0.5, 0.5, 0.5)
-        case .bottomRight: return FracRect(0.5, 0.5, 0.5, 0.5)
-        case .maximize: return FracRect(0, 0, 1, 1)
-        case .center: return FracRect(0.2, 0.15, 0.6, 0.7)
-        }
-    }
-
-    var title: LocalizedStringResource {
-        switch self {
-        case .leftHalf: return .shortcutLeftHalf
-        case .rightHalf: return .shortcutRightHalf
-        case .topHalf: return .shortcutTopHalf
-        case .bottomHalf: return .shortcutBottomHalf
-        case .topLeft: return .shortcutTopLeft
-        case .topRight: return .shortcutTopRight
-        case .bottomLeft: return .shortcutBottomLeft
-        case .bottomRight: return .shortcutBottomRight
-        case .maximize: return .shortcutMaximize
-        case .center: return .shortcutCenter
-        }
-    }
-}
-
 final class WindowManager {
 
     /// AX calls are synchronous IPC into the target app. Drag snapping makes
     /// one per mouse event, so a hung app must not take the cursor with it.
-    private static let axTimeout: Float = 0.25
+    static let axTimeout: Float = 0.25
 
     // MARK: - Coordinate conversion
 
@@ -104,113 +40,9 @@ final class WindowManager {
 
     // MARK: - AX helpers
 
-    var isTrusted: Bool { AXIsProcessTrusted() }
-
     /// Fires on the main queue after any window actually moves, whoever asked —
     /// hotkeys, drag snapping, workspace launches or the HTTP routes.
     var onDidPlace: (() -> Void)?
-
-    func promptForTrust() {
-        let opts = [kAXTrustedCheckOptionPrompt.takeUnretainedValue() as String: true] as CFDictionary
-        AXIsProcessTrustedWithOptions(opts)
-    }
-
-    private func appElement(_ pid: pid_t) -> AXUIElement {
-        let element = AXUIElementCreateApplication(pid)
-        AXUIElementSetMessagingTimeout(element, Self.axTimeout)
-        return element
-    }
-
-    private func axWindows(of pid: pid_t) -> [AXUIElement] {
-        var value: CFTypeRef?
-        guard AXUIElementCopyAttributeValue(appElement(pid), kAXWindowsAttribute as CFString, &value) == .success,
-              let windows = value as? [AXUIElement] else { return [] }
-        return windows.filter { (attr($0, kAXRoleAttribute) as? String) == kAXWindowRole }
-    }
-
-    private func attr(_ el: AXUIElement, _ name: String) -> CFTypeRef? {
-        var value: CFTypeRef?
-        guard AXUIElementCopyAttributeValue(el, name as CFString, &value) == .success else { return nil }
-        return value
-    }
-
-    func frame(ofWindow win: AXUIElement) -> CGRect? {
-        guard let posVal = attr(win, kAXPositionAttribute), let sizeVal = attr(win, kAXSizeAttribute) else { return nil }
-        var pos = CGPoint.zero
-        var size = CGSize.zero
-        guard AXValueGetValue(posVal as! AXValue, .cgPoint, &pos),
-              AXValueGetValue(sizeVal as! AXValue, .cgSize, &size) else { return nil }
-        return CGRect(origin: pos, size: size)
-    }
-
-    func focusedWindow(of app: NSRunningApplication) -> AXUIElement? {
-        let pid = app.processIdentifier
-        if let focused = attr(appElement(pid), kAXFocusedWindowAttribute),
-           CFGetTypeID(focused) == AXUIElementGetTypeID() {
-            return (focused as! AXUIElement)
-        }
-        return axWindows(of: pid).first { !isMinimized($0) }
-    }
-
-    private func isMinimized(_ win: AXUIElement) -> Bool {
-        (attr(win, kAXMinimizedAttribute) as? Bool) ?? false
-    }
-
-    private func title(of win: AXUIElement) -> String {
-        (attr(win, kAXTitleAttribute) as? String) ?? ""
-    }
-
-    /// The window under a point in AX space, or nil over the desktop. Walks up
-    /// from whatever element is there — a button, a text field — to the window
-    /// containing it.
-    func window(at point: CGPoint) -> AXUIElement? {
-        let system = AXUIElementCreateSystemWide()
-        AXUIElementSetMessagingTimeout(system, Self.axTimeout)
-        var element: AXUIElement?
-        guard AXUIElementCopyElementAtPosition(system, Float(point.x), Float(point.y), &element) == .success,
-              var current = element else { return nil }
-        // Ten levels is deeper than any real view hierarchy needs; the bound is
-        // there because a malformed AX tree can contain a cycle.
-        for _ in 0..<10 {
-            if (attr(current, kAXRoleAttribute) as? String) == kAXWindowRole { return current }
-            guard let parent = attr(current, kAXParentAttribute),
-                  CFGetTypeID(parent) == AXUIElementGetTypeID() else { return nil }
-            current = (parent as! AXUIElement)
-        }
-        return nil
-    }
-
-    func app(ofWindow win: AXUIElement) -> NSRunningApplication? {
-        var pid: pid_t = 0
-        guard AXUIElementGetPid(win, &pid) == .success else { return nil }
-        return NSRunningApplication(processIdentifier: pid)
-    }
-
-    /// Placement during a live drag. Announces nothing: the change bus would
-    /// otherwise fire on every mouse event, and the drop announces once.
-    func setFrame(_ rect: CGRect, ofWindow win: AXUIElement) {
-        setFrame(win, rect, announce: false)
-    }
-
-    @discardableResult
-    private func setFrame(_ win: AXUIElement, _ rect: CGRect, announce: Bool = true) -> Bool {
-        var pos = rect.origin
-        var size = rect.size
-        guard let posVal = AXValueCreate(.cgPoint, &pos),
-              let sizeVal = AXValueCreate(.cgSize, &size) else { return false }
-        // Position → size → position again: some apps clamp one against the other.
-        let movedTo = AXUIElementSetAttributeValue(win, kAXPositionAttribute as CFString, posVal)
-        let resized = AXUIElementSetAttributeValue(win, kAXSizeAttribute as CFString, sizeVal)
-        AXUIElementSetAttributeValue(win, kAXPositionAttribute as CFString, posVal)
-        guard movedTo == .success && resized == .success else { return false }
-        // Placement reaches here from hotkeys, drag snapping, workspace
-        // launches and the HTTP routes alike, so this is the one place that
-        // sees every window move. Launches run off the main queue.
-        if announce, let onDidPlace {
-            DispatchQueue.main.async(execute: onDidPlace)
-        }
-        return true
-    }
 
     private func axRect(for frac: FracRect, screenIndex: Int, in all: [ScreenInfo],
                         gap: CGFloat = 0) -> CGRect {
